requirements.txt
langgraph==0.2.33
langchain-core==0.3.23
langchain-openai==0.2.7
pypdf==4.3.1
python-dotenv==1.0.1
streamlit==1.41.1

